Renamed ranked1v1 to rating.  Refactored route.
